Update rand_core to 0.6 for 4.0.0 release.
`curve25519-dalek` exports the `rand_core` API as part of its API, so
changes to the `rand_core` (or `digest`) version require version bumps.
In the past, this has been an automatic major version bump as soon as
there was a new release, to prevent getting out of sync with the rest of
the ecosystem, and that should probably continue to create
`curve25519-dalek-ng` `4.0.0`.

Closes #2