Fix `--no-default-features` test warnings; consolidate CI jobs
Previously `cargo test --no-default-features` would succeed but with
warnings.

This commit fixes all of those warnings and tests
`--no-default-features` in CI to ensure that in perpetuity.

It moves the `tests-default-serde` into the `test` job which reduces the
total number of job runners needed and reuses intermediate artifacts
which should improve build times.

It removes `test-alloc-u32` as this case is already covered by the
`test` job when it runs `cargo test --target i686-unknown-linux-gnu` as
the `alloc` feature alone is now enabled by default.