Allow the fiat serial backend work with the simd backend
As discussed in 437 it is feasible to allow the fiat serial backend to
work with the simd backend.

This change makes it possible to use the fiat serial backend with the
simd backend with target_features `+avx2` and `+avx512ifma`.

This change introduces `cfg(curve25519_dalek_serial = "fiat")` so it can
work alonside with `cfg(curve25519_dalek_backend = "simd")`.

However for ergonomics reasons `cfg(curve25519_dalek_backend = "fiat")`
stil works without simd as before as does `curve25519_dalek_serial` alone.

It was also noted that the fiat serial backend requires MSRV 1.57.

The backend selection logic for this is now in build.rs.