ed25519-dalek: make `ExpandedSecretKey` fields private
Motivation can be found in this comment:

https://github.com/dalek-cryptography/ed25519-dalek/pull/293#pullrequestreview-1507115602

Replaces public fields with private fields + accessor methods.

Also adds an `ExpandedSecretKey::from_scalar_and_prefix` constructor
which makes it possible to construct `ExpandedSecretKey` using a
`Scalar` again, which is useful for protocols that derive the scalar
such as Ed25519-BIP32.

Additionally defines a `HashPrefix` type alias to make it semantically
clear which `[u8; 32]` is involved in type signatures.