Add support for the ff/group API to FieldElement
These changes are largely straightforward, with the note prior, the only
arithmetic supported prior was for borrows. Now that we've added support
for non-borrowed addition, the existing code is 'unnecessarily borrowing',
hence the added lint.

The larger question applicable is how this exposes FieldElement from several
different backends, assuming they're all correct for arbitrary usage. The
implementation of ConstantTimeEq has the following notes.

```rs
    /// Test equality between two `FieldElement`s.  Since the
    /// internal representation is not canonical, the field elements
    /// are normalized to wire format before comparison.
```

While that doesn't suggest the mathematical operations are incomplete, it
should be double-checked that no backend bounded the amount of operations
they're correct to on the presumption the curve formulas would never exceed so
many operations.