[WIP] curve: enable the `unreachable_pub` lint
In #850 I noticed there are a lot of places that internal types are
currently marked `pub` which is a bit confusing when trying to reason
about whether APIs allow for violating invariants.

This enables the `unreachable_pub` lint which can automatically check
for such situations. Most of them were corrected automatically using
`cargo fix`